登录修改

+++ src/IEManageSystem.Core/Entitys/Authorization/Users/AdminManager.cs
@@ -7,24 +7,41 @@ using System.Text;
 using System.Linq;
 using System.Threading.Tasks;
 using System.Linq.Expressions;
+using Abp.Domain.Repositories;
 
 namespace IEManageSystem.Entitys.Authorization.Users
 {
     public class AdminManager:IDomainService
     {
+        private IRepository<Role> _roleRepository { get; set; }
+
         private UserManager _userManager { get; set; }
 
         public AdminManager(
-            UserManager userManager)
+            UserManager userManager,
+            IRepository<Role> roleRepository)
         {
             _userManager = userManager;
+            _roleRepository = roleRepository;
         }
 
-        public User GetAdmin(int id) => _userManager.GetUser(id);
+        private IQueryable<User> _getAdmins(Expression<Func<User, object>>[] propertySelectors = null)
+        {
+            var role = _roleRepository.FirstOrDefault(e=>e.Name == Role.Admin.Name);
 
-        public IQueryable<User> GetAdmins() => _userManager.UserRepository.GetAll();
+            var admins = propertySelectors == null ? _userManager.UserRepository.GetAll() :
+                _userManager.UserRepository.GetAllIncluding(propertySelectors);
 
-        public IQueryable<User> GetAdminsIncluding(Expression<Func<User, object>>[] propertySelectors) => _userManager.UserRepository.GetAllIncluding(propertySelectors);
+            admins = admins.Where(e => e.UserRoles.Any(userRole => userRole.RoleId == role.Id));
+
+            return admins;
+        } 
+
+        public User GetAdmin(int id) => _getAdmins().FirstOrDefault(e => e.Id == id);
+
+        public IQueryable<User> GetAdmins() => _getAdmins();
+
+        public IQueryable<User> GetAdminsIncluding(Expression<Func<User, object>>[] propertySelectors) => _getAdmins(propertySelectors);
 
         public async Task<User> CreateAdmin(string userName, string password, string name = null, int? tenantId = null)
         {
@@ -37,12 +54,17 @@ namespace IEManageSystem.Entitys.Authorization.Users
 
         public void DeleteAdmin(int id)
         {
+            if (_getAdmins().FirstOrDefault(e => e.Id == id) == null)
+            {
+                throw new MessageException("找不到要删除的管理员");
+            }
+
             _userManager.DeleteUser(id);
         }
 
         public void UpdatePassword(int id, string password)
         {
-            var user = _userManager.GetUser(id);
+            var user = _getAdmins().FirstOrDefault(e => e.Id == id);
 
             if (user == null) {
                 throw new MessageException("找不到需要更新密码的管理员");
@@ -53,16 +75,30 @@ namespace IEManageSystem.Entitys.Authorization.Users
 
         public void UpdatePassword(User user, string password)
         {
+            if (_getAdmins().FirstOrDefault(e => e.Id == user.Id) == null) {
+                throw new MessageException("找不到需要更新密码的管理员");
+            }
+
             _userManager.UpdatePassword(user, password);
         }
 
         public void AddRole(User user, Role role)
         {
+            if (_getAdmins().FirstOrDefault(e => e.Id == user.Id) == null)
+            {
+                throw new MessageException("找不到需要添加权限的管理员");
+            }
+
             _userManager.AddUserRole(user, role);
         }
 
         public void Remove(User user, Role role)
         {
+            if (_getAdmins().FirstOrDefault(e => e.Id == user.Id) == null)
+            {
+                throw new MessageException("找不到需要移除权限的管理员");
+            }
+
             _userManager.RemoveRole(user, role);
         }
     }
